--story=1018804 --user=cug_zjr 推荐系统redis缓存

+++ recommend-server/src/main/java/com/xxxx/server/mongopojo/Rating.java
@@ -13,5 +13,5 @@ public class Rating  {
     private int userId;
     private int productId;
     private double score;
-    private long timestamp;
+    private int timestamp;
 }
